+++ MetricsManager/MetricsAgentTests/CpuMetricsControllerUnitTests.cs
@@ -1,0 +1,31 @@
+using MetricsAgent.Controllers;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using Xunit;
+
+namespace MetricsAgentTests
+{
+    public class CpuMetricsControllerUnitTests
+    {
+        private CpuMetricsController controller;
+
+        public CpuMetricsControllerUnitTests()
+        {
+            controller = new CpuMetricsController();
+        }
+
+        [Fact]
+        public void GetMetricsFromAgent_ReturnsOk()
+        {
+            //Arrange            
+            var fromTime = TimeSpan.FromSeconds(0);
+            var toTime = TimeSpan.FromSeconds(100);
+
+            //Act
+            var result = controller.GetMetricsFromAgent(fromTime, toTime);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+    }    
+}
