--- MetricsManager/MetricsAgentTests/UnitTest1.cs
@@ -1,127 +1,0 @@
-using MetricsAgent.Controllers;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using Xunit;
-
-namespace MetricsAgentTests
-{
-    public class CpuMetricsControllerUnitTests
-    {
-        private CpuMetricsController controller;
-
-        public CpuMetricsControllerUnitTests()
-        {
-            controller = new CpuMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange            
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class DotNetMetricsControllerUnitTests
-    {
-        private DotNetMetricsController controller;
-
-        public DotNetMetricsControllerUnitTests()
-        {
-            controller = new DotNetMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange            
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(fromTime, toTime);
-
-            // Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class HddMetricsControllerUnitTests
-    {
-        private HddMetricsController controller;
-
-        public HddMetricsControllerUnitTests()
-        {
-            controller = new HddMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }       
-    }
-
-    public class NetworkMetricsControllerUnitTests
-    {
-        private NetworkMetricsController controller;
-
-        public NetworkMetricsControllerUnitTests()
-        {
-            controller = new NetworkMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class RamMetricsControllerUnitTests
-    {
-        private RamMetricsController controller;
-
-        public RamMetricsControllerUnitTests()
-        {
-            controller = new RamMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }       
-    }
-}
