+++ MetricsManager/MetricsManagerTests/AgentControllerUnitTests.cs
@@ -1,0 +1,71 @@
+using MetricsManager;
+using MetricsManager.Controllers;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using Xunit;
+
+namespace MetricsManagerTests
+{
+    public class AgentControllerUnitTests
+    {
+        private AgentsController controller;        
+
+        public AgentControllerUnitTests()
+        {
+            controller = new AgentsController();            
+        }
+
+        [Fact]
+        public void RegisterAgent_ReturnsOk()
+        {
+            //Arrange
+            var agentInfo = new AgentInfo();
+
+            //Act
+            var result = controller.RegisterAgent(agentInfo);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+
+        [Fact]
+        public void EnableAgentById_ReturnsOk()
+        {
+            //Arrange
+            var agentId = 1;
+
+            //Act
+            var result = controller.EnableAgentById(agentId);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+
+        [Fact]
+        public void DisableAgentById_ReturnsOk()
+        {
+            //Arrange
+            var agentId = 1;
+
+            //Act
+            var result = controller.DisableAgentById(agentId);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+
+        [Fact]
+        public void GetAllAgents_ReturnsOk()
+        {
+            //Arrange
+            var agentInfo = new AgentInfo();
+
+            //Act
+            var result = controller.GetAllAgents(agentInfo);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+    }
+}
+
