+++ MetricsManager/MetricsManagerTests/CpuMetricsControllerUnitTests.cs
@@ -1,0 +1,47 @@
+﻿using MetricsManager;
+using MetricsManager.Controllers;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using Xunit;
+
+namespace MetricsManagerTests
+{
+    public class CpuMetricsControllerUnitTests
+    {
+        private CpuMetricsController controller;
+
+        public CpuMetricsControllerUnitTests()
+        {
+            controller = new CpuMetricsController();
+        }
+
+        [Fact]
+        public void GetMetricsFromAgent_ReturnsOk()
+        {
+            //Arrange
+            var agentId = 1;
+            var fromTime = TimeSpan.FromSeconds(0);
+            var toTime = TimeSpan.FromSeconds(100);
+
+            //Act
+            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+
+        [Fact]
+        public void GetMetricsFromAllCluster_ReturnsOk()
+        {
+            //Arrange
+            var fromTime = TimeSpan.FromSeconds(0);
+            var toTime = TimeSpan.FromSeconds(100);
+
+            //Act
+            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+    }
+}
