+++ MetricsManager/MetricsManagerTests/DotNetMetricsControllerUnitTests.cs
@@ -1,0 +1,48 @@
+﻿using MetricsManager;
+using MetricsManager.Controllers;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using Xunit;
+
+
+namespace MetricsManagerTests
+{
+    public class DotNetMetricsControllerUnitTests
+    {
+        private DotNetMetricsController controller;
+
+        public DotNetMetricsControllerUnitTests()
+        {
+            controller = new DotNetMetricsController();
+        }
+
+        [Fact]
+        public void GetMetricsFromAgent_ReturnsOk()
+        {
+            //Arrange
+            var agentId = 1;
+            var fromTime = TimeSpan.FromSeconds(0);
+            var toTime = TimeSpan.FromSeconds(100);
+
+            //Act
+            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
+
+            // Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+
+        [Fact]
+        public void GetMetricsFromAllCluster_ReturnsOk()
+        {
+            //Arrange
+            var fromTime = TimeSpan.FromSeconds(0);
+            var toTime = TimeSpan.FromSeconds(100);
+
+            //Act
+            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
+
+            //Assert
+            _ = Assert.IsAssignableFrom<IActionResult>(result);
+        }
+    }
+}
