--- MetricsManager/MetricsManagerTests/UnitTest1.cs
@@ -1,267 +1,0 @@
-using MetricsManager;
-using MetricsManager.Controllers;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using Xunit;
-
-namespace MetricsManagerTests
-{
-    public class AgentControllerUnitTests
-    {
-        private AgentsController controller;        
-
-        public AgentControllerUnitTests()
-        {
-            controller = new AgentsController();            
-        }
-
-        [Fact]
-        public void RegisterAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentInfo = new AgentInfo();
-
-            //Act
-            var result = controller.RegisterAgent(agentInfo);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void EnableAgentById_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-
-            //Act
-            var result = controller.EnableAgentById(agentId);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void DisableAgentById_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-
-            //Act
-            var result = controller.DisableAgentById(agentId);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetAllAgents_ReturnsOk()
-        {
-            //Arrange
-            var agentInfo = new AgentInfo();
-
-            //Act
-            var result = controller.GetAllAgents(agentInfo);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class CpuMetricsControllerUnitTests
-    {
-        private CpuMetricsController controller;
-
-        public CpuMetricsControllerUnitTests()
-        {
-            controller = new CpuMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetMetricsFromAllCluster_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class DotNetMetricsControllerUnitTests
-    {
-        private DotNetMetricsController controller;
-
-        public DotNetMetricsControllerUnitTests()
-        {
-            controller = new DotNetMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
-
-            // Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetMetricsFromAllCluster_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-
-        }
-    }
-
-    public class HddMetricsControllerUnitTests
-    {
-        private HddMetricsController controller;
-
-        public HddMetricsControllerUnitTests()
-        {
-            controller = new HddMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetMetricsFromAllCluster_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class NetworkMetricsControllerUnitTests
-    {
-        private NetworkMetricsController controller;
-
-        public NetworkMetricsControllerUnitTests()
-        {
-            controller = new NetworkMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetMetricsFromAllCluster_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-
-    public class RamMetricsControllerUnitTests
-    {
-        private RamMetricsController controller;
-
-        public RamMetricsControllerUnitTests()
-        {
-            controller = new RamMetricsController();
-        }
-
-        [Fact]
-        public void GetMetricsFromAgent_ReturnsOk()
-        {
-            //Arrange
-            var agentId = 1;
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAgent(agentId, fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-
-        [Fact]
-        public void GetMetricsFromAllCluster_ReturnsOk()
-        {
-            //Arrange
-            var fromTime = TimeSpan.FromSeconds(0);
-            var toTime = TimeSpan.FromSeconds(100);
-
-            //Act
-            var result = controller.GetMetricsFromAllCluster(fromTime, toTime);
-
-            //Assert
-            _ = Assert.IsAssignableFrom<IActionResult>(result);
-        }
-    }
-}
-
